requirements.txt
# Requirements for Hybrid Squeeze-Streaming ASR - Fixed Version

# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Configuration and data handling
pyyaml>=6.0
json5>=0.9.0

# Audio processing
librosa>=0.9.0
soundfile>=0.12.0

# Training utilities
tensorboard>=2.8.0
wandb>=0.12.0

# Evaluation metrics  
jiwer>=2.3.0
sacrebleu>=2.0.0

# Data processing
pandas>=1.3.0
tqdm>=4.62.0

# RNN-T Loss (optional - fallback to CTC if not available)
# warp_rnnt>=0.7.0

# Optional: For real-time transcription
pyaudio>=0.2.11

# Optional: For model optimization
onnx>=1.12.0
onnxruntime>=1.12.0

# Optional: For distributed training
accelerate>=0.16.0

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Additional dependencies
matplotlib>=3.5.0
seaborn>=0.11.0